rename file by changing the first letter
